Refactor service-related code and remove unused parameter-related code
