+++ src/app/controllers/__init__.py
@@ -1,11 +1,8 @@
 from app.api_services import api_services_controller
 
 from . import (
-    consumer_controllers,
     contact_us_controllers,
     my_favorites_controller,
-    parameters_controllers,
-    parameters_value_controllers,
     payment_controller,
     payment_profile_controllers,
     payment_status_controllers,
