+++ src/tests/controllers/__init__.py
@@ -1,5 +1,4 @@
 from . import (
-    test_parameter_controller,
     test_service_controllers,
     test_support_ticket_controllers,
     test_support_ticket_response_controllers,
